backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
psycopg2-binary==2.9.9
boto3==1.34.20
python-dotenv==1.0.0
gunicorn==21.2.0
pyjwt==2.8.0
pydantic==2.5.3
pydantic[email]==2.5.3
